Add widget includes
